refactor: create email feature
