+++ src/lib/consts/email.ts
@@ -1,0 +1,3 @@
+export const Email = {
+  FROM: 'move - studentische Unternehmensberatung e.V. <donotreply@transactional.consultingcontact.de>',
+} as const;
